requirements.txt
Selenium==4.0.0
playwright==1.20.0
pytest==7.0.0
python-dotenv==0.19.0